+++ CHANGES.md
@@ -17,6 +17,7 @@ Changes in 11.4
 - Simplify use of `$_SESSION['Back_PHP_SELF']` in MyReport.php, ClassSearchWidget.fnc.php, Search.inc.php & Export.php
 - Do NOT rely on SubmitButton() name param, use modfunc instead in DailyMenus.php, EditReportCardGrades.php, StudentAssignments.php, Calendar.php & Schools.php
 - Fix isPopup() for Calendar Events new modfunc detail_save & detail_delete in Warehouse.js
+- Use 'delete_ok' URL param instead of submit button name in Prompts.php
 
 Changes in 11.3.3
 -----------------